requirements.txt
Babel==2.7.0
boto==2.49.0
certifi==2019.6.16
chardet==3.0.4
Django==2.0.3
django-braces==1.13.0
django-extensions==2.2.1
django-formtools==2.1
django-model-utils==3.2.0
django-pagination==1.0.7
django-paypal==1.0.0
django-phone-field==1.8.0
django-phonenumber-field==3.0.1
django-private-chat==0.2.2
django-ses==0.8.12
filelock==3.0.12
future==0.18.0
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
phonenumbers==8.10.22
Pillow==6.1.0
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytz==2019.2
requests==2.22.0
six==1.12.0
stripe==2.35.1
toml==0.10.0
tox==3.14.1
urllib3==1.25.3
virtualenv==16.7.8
websockets==3.2
zipp==0.6.0
